backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "argon2-cffi>=23.1.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.1",
    "fastapi>=0.117.1",
    "httpx>=0.28.1",
    "huggingface-hub>=0.24.1",
    "mysql-connector-python>=9.4.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.11.9",
    "python-jose[cryptography]>=3.3.0",
    "pyjwt>=2.8.0",
    "pymysql>=1.1.2",
    "pytest>=8.4.2",
    "python-multipart>=0.0.20",
    "qrcode>=8.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]

[project.optional-dependencies]
dev = []
